Create Greenleaf servant leaders
With the publication of “The Servant as Leader” essay in 1970, followed by the book Servant Leadership in 1977, a different paradigm of leadership entered into the hearts and minds of people, organizations, and society. Robert K. Greenleaf proposed that service ought to be the distinguishing characteristic of leadership. Not only would it create a better, stronger society, but people themselves “would find greater joy in their lives if they raised the servant aspect of their leadership and built more serving institutions.”